+++ app/src/test/java/nl/tudelft/b_b_w/blockchain/BlockDataTest.java
@@ -18,7 +18,7 @@ public class BlockDataTest {
     private int sequenceNumber;
     private Hash previousHashChain;
     private Hash previousHashSender;
-    private int trustValue;
+    private double trustValue;
 
     /**
      * Setting up the test object
